plugin is working with cap 3
